Code cleanup for the item sheet
