+++ src/vue/components/parts/CodemirrorWrapper.vue
@@ -19,24 +19,26 @@
     if (game.modules.get('_CodeMirror')?.active && typeof CodeMirror !== undefined) {
       // Using a timeout for this is super hacky, but it's what we got.
       const element = textareaRef.value;
-      const editor = CodeMirror.fromTextArea(element, {
-        ...CodeMirror.userSettings,
-        mode: "javascript",
-        lineNumbers: true,
-        inputStyle: "contenteditable",
-        autofocus: false,
-        theme: game.settings.get("archmage", "nightmode") ? 'monokai' : 'default',
-        readOnly: element.hasAttribute('readonly')
-      });
-      // Update the textarea value so Foundry can see it.
-      editor.on('change', (instance) => {
-        rawValue.value = instance.getValue();
-      });
-
-      // Hacky, but refresh the editor content to fix quirks with CodeMirror + Vue.
-      setTimeout(() => {
-        editor.refresh();
-      }, 25);
+      if (element) {
+        const editor = CodeMirror.fromTextArea(element, {
+          ...CodeMirror.userSettings,
+          mode: "javascript",
+          lineNumbers: true,
+          inputStyle: "contenteditable",
+          autofocus: false,
+          theme: game.settings.get("archmage", "nightmode") ? 'monokai' : 'default',
+          readOnly: element.hasAttribute('readonly')
+        });
+        // Update the textarea value so Foundry can see it.
+        editor.on('change', (instance) => {
+          rawValue.value = instance.getValue();
+        });
+  
+        // Hacky, but refresh the editor content to fix quirks with CodeMirror + Vue.
+        setTimeout(() => {
+          editor.refresh();
+        }, 25);
+      }
     }
   });
 </script>
